backend/requirements.txt
azure-storage-blob
azure-identity
azure-cognitiveservices-speech
python-dotenv
pydub